ISSUE #2093 - FIxed blank page (circular reference)
