+++ frontend/routes/viewerCanvas/viewerCanvas.container.ts
@@ -24,7 +24,8 @@ import { selectPins as selectIssuePins } from '../../modules/issues';
 import { selectPins as selectMeasurementPins } from '../../modules/measurements';
 import { selectGISCoordinates, selectHasGISCoordinates } from '../../modules/model';
 import { selectPins as selectRiskPins } from '../../modules/risks';
-import { selectAllTransparencyOverrides, selectColorOverrides, TreeActions } from '../../modules/tree';
+import { TreeActions } from '../../modules/tree';
+import { selectAllTransparencyOverrides, selectColorOverrides } from '../../modules/viewerGui';
 import { withViewer } from '../../services/viewer/viewer';
 import { ViewerCanvas } from './viewerCanvas.component';
